All Battles!
I modified the battle function to return 0 when pokeA loses and 1 when it wins. 

I fixed a bug that was cumulatively reducing a pokmon's hp. Now when reading from the file of stats a copy of the pokemon info is made, instead of just referencing directly to the to the original stats dictionary.

I made 2 loops to simulate all the runs, but they are taking too long to run. Will leave it running tomorrow during TA.

The last line of code is a potential way to find out the pokemon with highest number of victories.